+++ src/modules/edc-demo/components/catalog-browser-fetch-detail-dialog/catalog-browser-fetch-detail-dialog.data.ts
@@ -1,0 +1,7 @@
+import {Observable} from 'rxjs';
+import {CatalogBrowserPageData} from '../catalog-browser/catalog-browser-page.data';
+
+export interface CatalogBrowserFetchDetailDialogData {
+  data$: Observable<CatalogBrowserPageData>;
+  refresh: () => void;
+}